Created Spotify API Object with Spotipy
Created the object needed to access Spotify's API. Also added the ability for the program to use environment variables to since they're needed for the api. See the guide in the drive to set up the .env file.

Note: In the requirements, we might need to keep idna at 2.10 instead of 3.10 because of other packages.