Implemented SpotifySDK, added Play/Pause, Next/Previous track, and Volume controls, & rearranged Views structure
This commit implements the SpotifySDK in the _footer.html script tags and applies logic to the Play/Pause, Next/Previous track, and Volume controls. On the frontend, these controls simply ajax to access the backend while also providing the device ID for the browser. SpotifySDK was implemented to allow for playing the music in-browser, and also its the one that generates our device id. To make the pathing a little more clean for the ajax calls, I moved views.py into a directory called views which is now split into pages.py and spotify_api.py. These get imported via the __init__.py file in the views directory to allow for easier pathing in urls.py. In spotpy_api.py, I moved the redirectToHome function here and implemented setPlayback, setVolume, nextTrack, and previousTrack which take in the ajax'd data from the frontend and apply them to spotiPy's API calls. At the moment, you have to hit the play button twice to get it started. I still need to figure out that bug.