Added track info to player & moved button events to separate functions
This commit adds the basic track info to the bottom left of the player, which is the song name and artist(s). These get updated whenever the SpotifySDK state gets updated. I store this state in the deviceState variable to make it accessible elsewhere. I also moved all the button events into their own functions to be called elsewhere when needed. Finally, I moved the SpotifySDK import into the webplayer.html skeleton file.