Added Shuffle control, toggleShuffle event, & setShuffle backend function
I added two buttons to this commit, the shuffle and repeat controls. The repeat control has no functionality set. The shuffle control triggers toggleShuffleEvent which utilizes an isShuffled variable to make ajax calls to the setShuffle function in the backend. The backend will take this request, determine if shuffle_status was set to enabled or disabled, and then set the device's shuffle status. In the state update event for the SpotifySDK, I have an initial check to see if the Shuffle value is unset. If it is, I grab the device's shuffle status via SpotifySDK and update it there. The reason I don't always update the text here is because spam-clicking the button would cause the visual updates to delay and even show incorrect results at times. Finally, I set a height value for the buttons to fix a flex-box CSS issue