Added Repeat control, toggleRepeat event, & setRepeat backend function
During the Spotify state update, I initially check to see if the repeat control has default values. If so, I update repeatStatus and set the control's text to either track, context, or off. The toggleRepeat function will increment repeatStatus and limit it from 0 to 2. I then ajax to setRepeat in the backend by passing the result of repeatStatus. The setRepeat backend function will take the repeatStatus value and search for it in the in a states dictionary and apply it to the SpotiPy function. If successful, we'll return that value to the frontend to update the control text. If it failed, then on the frontend we'll decrement repeatState again. Finally, I tried to clean up the setShuffle function in the backend.