Added Play Playlist button to auto-generated songs page
This commit allows the user to navigate to the song list for a specific playlist and press 'PLAY PLAYLIST'. I temporarily added a click event to the webplayer.html script tags to utilize the Django URL syntax. It will simply ajax to the setPlayback endpoint. I modified this setPlayback endpoint to also accept an optional context_uri variable to pass into sp.start_playback. I also made it so that if no status is provided, playback will be set to play automatically. Finally for the endpoints, I modified the call to createSongList which will pass in a third parameter: the playlist id. In createSongList, I check to see if context_uri and type are passed, and if so, I modify the context_uri to include 'spotify:TYPE:ID'. This is the necessary context_uri needed to make API calls for setting playlists as the current context. The first section of the HTML string contains the newly created button that has a URI data parameter to use in the associated ajax call.