Refactored togglePlayback() and how Play/Pause visuals update
Since playing a playlist is designed to automatically resume playback, we needed a way to update the play/pause visual. I now have it set so the play/pause image gets set when SpotifySDK updates its context variable. This will also set the isPlaying variable which we use in togglePlayback(). Now togglePlayback() is a single ajax command that passes a playbackRequest variable into the status parameter. I also had to replace a line from Ryan's earlier code back into this refactored version of the function.