Created HTML pages for User Collections
Created a general HTML page that display user collections (artist, album, playlist, podcast).

create_html.py has been updated with comments saying which functions are now deprecated with the introduction of these HTML pages. From here on out, create_html.py will eventually be deleted and we will use these HTML pages instead.

Before, we were using strings to create HTML since we didn't have a solution to render an HTML page dynamically and pass it back to ajax. This has been solved with the use of a Django function called render_to_string which returns a rendered page as a string.

The HTML pages solve the issue of what happens when a user manually goes to a user collection in the address bar. Now, whether by clicking the link on the page or manually typing the url, the page will be displayed correctly.

script.js has been updated to work with the new HTML page. We will have one event listener for user collections and one event listener for navigating to an item's page. This should not be a problem since most of the information is processed in the backend and ajax mainly inserts the HTML string returned.