Added non-functional context menu on right click
This commit adds in the basics of a right-click enabled menu that changes button options depending on the context. There is extremely minimal styling here, so this will be for a later commit. Here I added the mouse-menu to the page and added a contextmenu event to the document that will populate the menu with options related to the clicked item. To keep this consistent, I've added a few general classes we can use on elements of our page, such as Artist, Playlist, Album, Podcast, and Song. Now the ContentItem classified elements from collectionItems.html will be given one of these classes depending on their passed type. Also, I added these new classes to the track information in the bottom-left of the playbar. Currently these buttons have no functionality, but they will in a later commit. After populating the menu, we record the mouse location and rearrange the menu's position depending on how close to the edge it is. I also added a click event to the document that will hide this mouse-menu if the user clicks anywhere away from the menu after its opened. Right-click will also hide the menu and then re-open it if applicable in the contextmenu event. And finally, the most minimal change is that I added a hover effect to the song list to easily differentiate if you're right-clicking on a song or not.