Added missing data-uri attributes, made playbar artists into links, and hid the playbar until it's ready
For the context menu, we'll need to extract URIs in order to perform various Spotify functions. Therefore, I added data-uri attributes to each element that has access to an associated Spotify URI. The SpotifySDK provides the full URI, so for consistency with how we handled the other URIs, I only extracted the ID part. Also, even though the ContentItem IDs are set as the URI values on the collectionItems.html page, this could cause a unique ID problem in the future. Therefore, I just repasted that information into the data-uri tags. In the future we'll need to remove that existing id tag on all ContentItems. On this page I also added the Artist class to the ContentArtist divs and attached the data-uri attribute, too. We'll use the class to differentiate what API parameter to pass, (such as spotify:TYPE:ID). I also changed the single trackArtist p tags into a span container that gets anchor tags added per artist. Their hrefs are also set to mimic how we set hrefs on the Song List pages. Finally, the footer is now hidden by default and gets animated onto the page once it's ready to be used.