Albums and Related Artists shown for Artist Page
Started writing the artist page. Very minimal as of now, no new styling.

The html page is divided into 4 different sections: the header, the interaction bar, the selection bar with what tabs you want to view, and the tab contents.

Each section will be filled as we go.

Functionality to view the artist's albums and related artists have been implemented. The selection bar is not finished, but provides enough functionality to test each tab.

Top songs have been semi-implemented. The information is there and is on the page as just text. Will be finished when the song list creation html is provided by another team member.

Create_html.py will probably turn into a helper file as new html pages are created.