Refactored all URIs to utilize full Spotify URI, made songs playable through Song List, added PlayRequest class to trigger context playing, & minor CSS changes
This is a huge commit that probably needs some thorough look through the code to fully capture everything. First off, I refactored all of the previous commit URI logic to store the entire Spotify URI instead of just the song IDs. Also, I removed existing HTML ids that were set to the song ID to avoid potential duplicate-ID problems in the future. I also added the PlayRequest class that, when clicked, will gather the object's data-uri and data-parent-uri to utilize for context and offset URIs if applicable. Therefore, an object we deem a PlayRequest button must have a data-uri if it's for a playlist, album, or artist. If it's a track, then it's data-uri will be the track URI and it's data-parent-uri will be its context (playlist, album, or artist). This is because you have to provide a context when playing individual songs so it knows where to resume if you disable track loop. The PlayRequest click functionality will then ajax to setPlayback and pass a 'play' status alongside the context_uri and offset_uri if its available. If it succeeds, then it'll hide the mouse-menu if it hasn't been hidden already. The mouse-menu functionality also adds this PlayRequest class to the Play Playlist/Album/Artist buttons when right-clicking. This commit also adds a hover functionality to the Song List where the song number will turn into a play button (with the PlayRequest class) to play that single track.  All of this PlayRequest implementation warranted changes in the backend, too. Now setPlayback() will check for an offset URI and pass it to SpotiPy if necessary. In the playlist() function, I included the songURI to the info object and appended 'spotify:playlist:' to the previously passed playlistID since we only get the ID from the URL, not the rest of the required URI syntax. In create_html.py's createSongList(), I removed the context URI and type string generator. (This is also where I made those changes to classes and added data-uri attributes). And finally, for CSS changes, I removed the hidden footer functionality since it would eventually cause miscalculations for the page scroll. I also removed the focus styling on the footer buttons since clicking one would trigger focus and thus leave a fade on the button.