Refactored toggleShuffle() and updated Shuffle visuals from SpotifySDK
The toggleShuffle() function was simplified to avoid repetitive code and it no longer sets the isShuffled value. That value is now determined by the SpotifySDK context which is also where we will update the visuals of the button. Like toggleRepeat, this function now has a timeout function to avoid spam clicking.