Fixed data-uri bug for Song List artists & fixed caching bug for the playbar's album data-uri
I forgot to change the Song List's artist data-uris to the full spotify:artist:ID version, so I included that in this commit. Also, attempting to "Play Album" from the playbar's album image was causing problems because jQuery's data() function caches the initial result. I replaced it with attr() to fix this problem.