Implemented basic sidebar queue functionality and styling
This commit serves as the foundation for the sidebar queue. The sidebar html file implements this section beneath the navigation buttons (where I also re-arranged and commented out unused links). The queue also contains a settings button (which we can utilize later) and a toggle arrow for showing and hiding the queue.  To stylize this queue, I added a bunch of classes to this file's style tags. However, for functionality, I implemented this in the footer file. This is because the queue is populated by information from the SpotifySDK. Eventually we should move all of this stuff to separate JavaScript files. I moved the method of converting a song's artist into clickable links to its own function named getArtistAsLinks(). In the SpotifySDK context change event, I set all of the queue items depending on if the previous & next tracks are available. If not, I hide them from the queue. We can eventually replace this and just leave unavailable data as blank or faded out. My future commits will need to fix the overlapping text problem  and apply the same changes to the footer if possible.