Adjustments to Queue and Footer CSS
The queue now has a shadow on the top-most element to show the end of the queue. Also, the current song has a larger image now. For all the queue items, margins were reduced and song titles bolded. On both the footer and the queue, I temporarily cut off text that exceeds the set width. This now means in the footer, the left and right sides are set at 300px and the center fills whatever is left. Finally, I rephrased the alt text of the footer album cover.