Replaced footer queue button with SpotifySDK status & fixed some footer bugs
Since the queue is on the sidebar now, I removed the unnecessary queue button from the footer. It has now been replaced by a little signal indicator that tells if the user is currently connected to SpotifySDK. If not, the signal will show no connection. Clicking this signal will force a playback transfer request to return control to Modivibe. I also fixed an image scaling problem on the footer's album art and an incorrect play button icon when loading the page.