Artist Top Songs can now be played individually & Bug fixes to Queue logic
Before we weren't able to play artist top songs since you can't offset into an artist context. I'm not sure why, but it seems to be a Spotify API restriction.  To overcome this situation, I modified the setPlayback endpoint to extract a contextType from the contextURI. If the type is artist, then we'll grab the artistTopTracks and extract their URIs to set as the current context. To avoid repeats, we remove the offset URI from this list (the offset URI is just the song the user clicked play on). After that I pass the topTrackURIs to the uris parameter of sp.start_playback function  and offset on the song the user clicked played on.

I also made some adjustments to the Queue logic that was causing some bugs when certain information was unavailable and not correctly hiding.