Fixed content loading for artists by adding ItemLink class to getArtistAsLinks() and Song List artists
We forgot to add the ItemLink class to these anchor tags, so clicking them would cause the page to redirect instead of ajaxing and updating the content area dynamically. Adding the class to the getArtistAsLinks() function affected the links in the footer and the queue.  Adding the class to the Song List page now loads the content properly, but we found another bug where these links can point to inaccurate artistID href values despite different artist names.