Implemented Follow Artist logic on Artist Page
First I created a click event for FollowRequest class objects that takes their associated data-uri and data-following attributes to determine what to ajax. Data-uri is just the artist URI and data-following is either true or false (which we will update on page load and after successfully toggling follow). With this information, we ajax to the toggleFollow endpoint in spotify_api.py.

The backend toggleFollow function takes the artist URI and follow status. Since SpotiPy's follow/unfollow functions requires a list of ids, we split the URI to get its ID and place it in a single-item list. The follow or unfollow function we call is dependent on the follow parameter. We either return True or False to show process was successful. Back to the ajax success function, we then call updateFollowStatus()

The updateFollowStatus() function finds all FollowRequest buttons and creates an affectedURIs list. Then, for each follow button, we grab its artist URI and check if it's already been processed. If so, we'll skip it. We then ajax to the isFollowing endpoint in spotify_api.py.

The isFollowing backend function just takes the artist URI, puts it in a single-item list, and returns a boolean list of whether or not the user follows the artist. If this succeeds, we'll return a json response with the artist URI and the following status.

Back to the success function from the updateFollowStatus ajax call, we check to see if a non-False response was returned (implying a json was returned). If so, we parse the json and find all FollowRequest objects that have the same data-uri as the artist URI we returned. We then set their data-following attribute to the following status we returned. Finally, we update the text to display the correct material icon logo.

The Artist Page implements this functionality by applying FollowRequest to a button that contains a data-uri (with the artist URI) and a data-following attribute. These attributes are set from create_html.py's getArtistHeaderInfo() function. The text (for the material icon designs) is also set depending on the value we passed into data-following.

Later commits will need to apply the FollowRequest button to the right-click context menu.