Added recommendations.html, added Queue.css, added getCookie() function, started recommendation endpoint implementation
This commit is just a progress save check. First I added the recommendations.html file and moved the sliders Alex made on the Settings page to the new Recommendation page. I'll eventually need to remove those sliders from the settings page, too. Also, I added new sliders to use all the features available (except Mode, Duration, and Time Signature). The getRecommendations endpoint was also created and currently is accepting the parameters correctly and taking a manually set artist for testing. My next commits will need to set the Reference recommendation logic.

The Queue.css file is just the CSS used in the sidebar. I moved it to its own file since I'm reusing the styling on the Recommendation list.

In utl.js, I added the getCookie() function recommended by Django to use a better method of collecting the csrftoken. I'll need to replace all ajax references to this with the new const variable.