Artist Recommendation initial implementation
First, I added the RecommendationRequest class click event where, when clicked, it'll pass the data-uri of the object to the recommendation backend function and load the page with the passed data.

The recommendations backend page load function will take this URI and call getContextURIInfo (newly added function) to pass referenceInfo and referenceString into the page context.

getContextURIInfo is a helper function that returns a list of various context information like name, uri, picture, etc.

Recommendations.html uses the context's referenceInfo item to set header information and referenceString to pass in the ajax for generating recommendations. The endpoint: getRecommendations()

The getRecommendations() endpoint takes the referenceString and de-stringifies it (to make it a json). After that, depending on the type passed, we extract certain seed values for genres, tracks, and artists. For now, only the artist work is implemented. We then pass that into the SpotiPy recommendation API call.

I also changed the logic for the slider functionality because it was throwing errors due to our dynamic page content system.

Finally, I added a Recommend Songs button on the artist right-click menu that uses the RecommendationRequest class.