Added in Album Recommendation functionality
First I added the RecommendationRequest class to a Recommend Songs right click menu for Albums.

Second I added logic to getContextURIInfo for albums. If an album has no assigned genres, we'll find its artist and take their genres instead. Also this referenceData object passes artistURI and artistName.

Third I added logic to getRecommendations for albums. For albums, we extract 1 artist URI, 1-3 genres, and a random track from the album.

Fourth, I added Django if statements to recommendations.html that differentiate the headers depending on the recommendation context. Artists need to be converted to links, so I added a jQuery function that uses the data-uri and data-name to create one artist to pass into getArtistAsLinks().