Added in Playlist Recommendation functionality
First I added the RecommendationRequest button to the right-click menu for Playlists.

Second I updated getcontextURIInfo to handle playlists. If given a playlist, we find 15 non-local tracks and append it to the referenceData object.

Third, I Updated getRecommendations to handle playlists. All this does is set the tracks seed to 5 full tracks. I also updated the genres setting for artists and albums to make more sense.

Fourth, I added a Django if statement for playlists to the recommendations.html file that only displays the Playlist name.