Added in Song Recommendation functionality & fixed Queue Song data-uri bug
First, I added the RecommendationRequest button to the right-click menu for songs.

Second, I added track logic for getContextURIInfo. We pass the artistURI, name, and tracks of the container album into referenceData.

Third, I added track logic to getRecommendations that just passes the artistURI to artists and 4 reference tracks into tracks.

Fourth, I added a Django if statement for tracks that display the track name and the artist. I also added #trackArtistName as a selector for the jQuery function that gets the artist name as a link.

Fifth, I fixed a bug in the Queue where the data-uri wasn't being applied.