Applied Track Features on Song Recommendation, added slider descriptions, & adjusted CSS of page
First, I extracted audio features in the getContextURIInfo function so that it gets passed into the context variable when loading the recommendations page. These get used to populate the slider values if available.

Second, I set the values of each slider depending on if they exist in the given context.

Third, I added help buttons next to each slider that, when clicked, shows a description of each feature.

Fourth, I adjusted the CSS throughout the page for most consistent visuals.